chore: replace op-reth OpReceipt with op-alloy's
