+++ crates/optimism/primitives/src/lib.rs
@@ -11,6 +11,17 @@
 #![allow(unused)]
 extern crate alloc;
 
+#[cfg(any(test, feature = "arbitrary"))]
+use arbitrary as _;
+#[cfg(feature = "reth-codec")]
+use bytes as _;
+#[cfg(feature = "reth-codec")]
+use modular_bitfield as _;
+#[cfg(any(test, feature = "reth-codec"))]
+use reth_codecs as _;
+#[cfg(feature = "reth-codec")]
+use reth_zstd_compressors as _;
+
 pub mod bedrock;
 
 pub mod predeploys;
@@ -20,7 +31,8 @@ pub mod transaction;
 pub use transaction::*;
 
 mod receipt;
-pub use receipt::{DepositReceipt, OpReceipt};
+pub use op_alloy_consensus::OpReceipt;
+pub use receipt::DepositReceipt;
 
 /// Optimism-specific block type.
 pub type OpBlock = alloy_consensus::Block<OpTransactionSigned>;
@@ -44,6 +56,6 @@ impl reth_primitives_traits::NodePrimitives for OpPrimitives {
 /// Bincode-compatible serde implementations.
 #[cfg(feature = "serde-bincode-compat")]
 pub mod serde_bincode_compat {
-    pub use super::receipt::serde_bincode_compat::*;
-    pub use op_alloy_consensus::serde_bincode_compat::*;
+    pub(crate) use super::receipt::serde_bincode_compat::*;
+    pub(crate) use op_alloy_consensus::serde_bincode_compat::*;
 }
