+++ crates/primitives-traits/src/serde_bincode_compat.rs
@@ -345,4 +345,17 @@ mod block_bincode {
             repr.into()
         }
     }
+
+    #[cfg(feature = "op")]
+    impl super::SerdeBincodeCompat for op_alloy_consensus::OpReceipt {
+        type BincodeRepr<'a> = op_alloy_consensus::serde_bincode_compat::OpReceipt<'a>;
+
+        fn as_repr(&self) -> Self::BincodeRepr<'_> {
+            self.into()
+        }
+
+        fn from_repr(repr: Self::BincodeRepr<'_>) -> Self {
+            repr.into()
+        }
+    }
 }
