+++ crates/primitives-traits/src/size.rs
@@ -148,6 +148,15 @@ mod op {
         }
     }
 
+    impl InMemorySize for op_alloy_consensus::OpReceipt {
+        fn size(&self) -> usize {
+            match self {
+                Self::Legacy(receipt) | Self::Eip2930(receipt) | Self::Eip1559(receipt) | Self::Eip7702(receipt) => receipt.size(),
+                Self::Deposit(receipt) => receipt.size(),
+            }
+        }
+    }
+
     impl InMemorySize for op_alloy_consensus::OpTypedTransaction {
         fn size(&self) -> usize {
             match self {
